Added supports functionality

+++ TrayReportApp.Models/Support.cs
@@ -8,7 +8,7 @@ namespace TrayReportApp.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
-        public int Weight { get; set; }
-        public int Distance { get; set; }
+        public double Weight { get; set; }
+        public double Distance { get; set; }
     }
 }
