+++ TrayReportApp.Services/Models/SupportServiceModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
-        public int Weight { get; set; }
-        public int Distance { get; set; }
+        public double Weight { get; set; }
+        public double Distance { get; set; }
     }
 }
